support clipboard

+++ tools/idl-gen/src/lib.rs
@@ -7,11 +7,11 @@ use std::net::IpAddr;
 pub use gen::volo_gen::*;
 use volo::FastStr;
 
-impl From<(IpAddr, u32)> for info::Host {
-    fn from(val: (IpAddr, u32)) -> Self {
+impl From<(IpAddr, u16)> for info::Host {
+    fn from(val: (IpAddr, u16)) -> Self {
         info::Host {
             ip_addr: FastStr::new(val.0.to_string()),
-            port: val.1,
+            port: val.1 as u32,
         }
     }
 }
